chore: separating GeoLayer implementation to a helper #437
